+++ pkg/operator/controller/status_reconciler/status_reconciler_test.go
@@ -79,7 +79,7 @@ var _ = Describe("Status Controller", func() {
 				if condition.Type == string(kaiv1.ConditionTypeReconciling) &&
 					condition.Status == metav1.ConditionTrue &&
 					condition.ObservedGeneration == kaiConfig.Generation &&
-					condition.Reason == string(kaiv1.Reconciled) {
+					condition.Reason == string(kaiv1.Reconciling) {
 					found = true
 				}
 			}
@@ -116,6 +116,8 @@ var _ = Describe("Status Controller", func() {
 			isAvailable = isAvailable && !isAvailableErr
 			Expect(checkIsDeployed(object.GetConditions())).To(Equal(isDeployed))
 			Expect(checkIsAvailable(object.GetConditions())).To(Equal(isAvailable))
+			Expect(checkIsReady(object.GetConditions())).To(Equal(isAvailable))
+			Expect(checkIsReconciling(object.GetConditions())).To(BeFalse())
 		},
 			Entry("kai config - no errors, all set", func() objectWithConditions { return &KAIConfigWithStatusWrapper{Config: kaiConfig} }, false, true, false, true),
 			Entry("kai config - no errors, deployed not avaialbe", func() objectWithConditions { return &KAIConfigWithStatusWrapper{Config: kaiConfig} }, false, true, false, false),
@@ -146,6 +148,22 @@ func checkIsAvailable(conditions []metav1.Condition) bool {
 		return false
 	}
 	return isAvailableCondition.Status == metav1.ConditionTrue
+}
+
+func checkIsReady(conditions []metav1.Condition) bool {
+	isReadyCondition := getConditionByType(conditions, string(kaiv1.ConditionTypeReady))
+	if isReadyCondition == nil {
+		return false
+	}
+	return isReadyCondition.Status == metav1.ConditionTrue
+}
+
+func checkIsReconciling(conditions []metav1.Condition) bool {
+	isReconcilingCondition := getConditionByType(conditions, string(kaiv1.ConditionTypeReconciling))
+	if isReconcilingCondition == nil {
+		return false
+	}
+	return isReconcilingCondition.Status == metav1.ConditionTrue
 }
 
 func getConditionByType(conditions []metav1.Condition, conditionType string) *metav1.Condition {